backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.29.0
python-jose==3.3.0
passlib==1.7.4
pydantic==2.6.4
pydantic-settings==2.2.1 
motor==3.3.2
python-multipart==0.0.9
email-validator==2.1.1
pymongo==4.6.3

